move some more formulas
