+++ uds/translator/data_record_definitions/__init__.py
@@ -17,7 +17,7 @@ from .conditional import (
     CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_CLIENT,
     CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_SERVER,
     CONDITIONAL_OPTIONAL_NEEDED_ADDITIONAL_PARAMETER,
-    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,
+    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,SCALING_DATA_RECORDS,
     CONDITIONAL_PROOF_OF_OWNERSHIP_CLIENT,
     CONDITIONAL_PROOF_OF_OWNERSHIP_SERVER,
     DTC_SNAPSHOT_RECORDS_LIST_2013,
@@ -139,7 +139,6 @@ from .other import (
     POWER_DOWN_TIME,
     PROOF_OF_OWNERSHIP_CLIENT_LENGTH,
     PROOF_OF_OWNERSHIP_SERVER_LENGTH,
-    SCALING_DATA_RECORDS,
     SECURITY_DATA_REQUEST_RECORD,
     SECURITY_DATA_RESPONSE_RECORD,
     SERVICE_TO_RESPOND,
