+++ uds/translator/data_record_definitions/conditional.py
@@ -28,9 +28,9 @@ __all__ = [
     "EVENT_TYPE_RECORD_09_2020", "CONDITIONAL_EVENT_TYPE_RECORD_09_2020"
 ]
 
-from uds.utilities import REPEATED_DATA_RECORDS_NUMBER
+from uds.utilities import EXPONENT_BIT_LENGTH, MANTISSA_BIT_LENGTH, REPEATED_DATA_RECORDS_NUMBER
 
-from ..data_record import ConditionalFormulaDataRecord, RawDataRecord
+from ..data_record import ConditionalFormulaDataRecord, CustomFormulaDataRecord, RawDataRecord
 from .dtc import (
     DTC_EXTENDED_DATA_RECORDS_DATA_LIST,
     DTC_STORED_DATA_RECORD_NUMBERS_LIST,
@@ -50,11 +50,13 @@ from .formula import (
     get_event_type_record_07_2013,
     get_event_type_record_07_2020,
     get_event_type_record_09_2020,
-    get_formula_for_raw_data_record_with_length,
+    get_formula_raw_data_record_with_length,
+    get_formula_scaling_byte_extension,
     get_max_number_of_block_length,
     get_max_number_of_block_length_file_transfer,
     get_memory_size_and_memory_address,
 )
+from .other import SCALING_BYTE_LENGTH, SCALING_BYTE_TYPE
 
 DTC_DIDS_RECORDS_LIST_2013 = [
     ConditionalFormulaDataRecord(formula=get_did_records_formula_2013(record_number + 1))
@@ -106,70 +108,89 @@ CONDITIONAL_MEMORY_ADDRESS_AND_SIZE = ConditionalFormulaDataRecord(formula=get_m
 CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH = ConditionalFormulaDataRecord(formula=get_max_number_of_block_length)
 """Definition of conditional `maxNumberOfBlockLength` Data Record."""
 
+# SID 0x24
+
+SCALING_BYTES_LIST = [RawDataRecord(name=f"scalingByte#{index + 1}",
+                                    children=(SCALING_BYTE_TYPE, SCALING_BYTE_LENGTH),
+                                    length=8,
+                                    min_occurrences=1 if index == 0 else 0,
+                                    max_occurrences=1)
+                      for index in range(REPEATED_DATA_RECORDS_NUMBER)]
+"""Collection of `scalingByte` Data Records."""
+
+SCALING_BYTES_EXTENSIONS_LIST = [ConditionalFormulaDataRecord(formula=get_formula_scaling_byte_extension(index + 1))
+                                 for index in range(REPEATED_DATA_RECORDS_NUMBER)]
+"""Collection of `scalingByteExtension` Data Records."""
+
+SCALING_DATA_RECORDS = [item for scaling_data_records in zip(SCALING_BYTES_LIST,
+                                                             SCALING_BYTES_EXTENSIONS_LIST)
+                        for item in scaling_data_records]
+"""Collection of `scalingByte` and `scalingByteExtension` Data Records."""
+
 # SID 0x29
 
 CONDITIONAL_CERTIFICATE_CLIENT = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="certificateClient",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="certificateClient",
+                                                    accept_zero_length=False))
 """Definition of conditional `certificateClient` Data Record."""
 
 CONDITIONAL_CERTIFICATE_SERVER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="certificateServer",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="certificateServer",
+                                                    accept_zero_length=False))
 """Definition of conditional `certificateServer` Data Record."""
 
 CONDITIONAL_CERTIFICATE_DATA = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="certificateData",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="certificateData",
+                                                    accept_zero_length=False))
 """Definition of conditional `certificateData` Data Record."""
 
 CONDITIONAL_CHALLENGE_CLIENT = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="challengeClient",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="challengeClient",
+                                                    accept_zero_length=False))
 """Definition of conditional `challengeClient` Data Record."""
 CONDITIONAL_OPTIONAL_CHALLENGE_CLIENT = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="challengeClient",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="challengeClient",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `challengeClient` Data Record."""
 
 CONDITIONAL_CHALLENGE_SERVER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="challengeServer",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="challengeServer",
+                                                    accept_zero_length=False))
 """Definition of conditional `challengeServer` Data Record."""
 
 CONDITIONAL_PROOF_OF_OWNERSHIP_CLIENT = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="proofOfOwnershipClient",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="proofOfOwnershipClient",
+                                                    accept_zero_length=False))
 """Definition of conditional `proofOfOwnershipClient` Data Record."""
 
 CONDITIONAL_PROOF_OF_OWNERSHIP_SERVER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="proofOfOwnershipServer",
-                                                        accept_zero_length=False))
+    formula=get_formula_raw_data_record_with_length(data_record_name="proofOfOwnershipServer",
+                                                    accept_zero_length=False))
 """Definition of conditional `proofOfOwnershipServer` Data Record."""
 
 CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_CLIENT = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="ephemeralPublicKeyClient",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="ephemeralPublicKeyClient",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `ephemeralPublicKeyClient` Data Record."""
 
 CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_SERVER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="ephemeralPublicKeyServer",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="ephemeralPublicKeyServer",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `ephemeralPublicKeyServer` Data Record."""
 
 CONDITIONAL_OPTIONAL_NEEDED_ADDITIONAL_PARAMETER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="neededAdditionalParameter",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="neededAdditionalParameter",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `neededAdditionalParameter` Data Record."""
 
 CONDITIONAL_OPTIONAL_ADDITIONAL_PARAMETER = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="additionalParameter",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="additionalParameter",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `additionalParameter` Data Record."""
 
 CONDITIONAL_OPTIONAL_SESSION_KEY_INFO = ConditionalFormulaDataRecord(
-    formula=get_formula_for_raw_data_record_with_length(data_record_name="sessionKeyInfo",
-                                                        accept_zero_length=True))
+    formula=get_formula_raw_data_record_with_length(data_record_name="sessionKeyInfo",
+                                                    accept_zero_length=True))
 """Definition of optional conditional `sessionKeyInfo` Data Record."""
 
 # SID 0x2C
